pyedb hierarchy terminal refactoring
