+++ src/pyedb/dotnet/edb_core/cell/layout_obj.py
@@ -126,7 +126,7 @@ class Connectable(LayoutObj):
         -------
         :class:`dotnet.edb_core.edb_data.nets_data.EDBComponent`
         """
-        from pyedb.dotnet.edb_core.edb_data.components_data import EDBComponent
+        from pyedb.dotnet.edb_core.hierarchy.component import EDBComponent
 
         edb_comp = self._edb_object.GetComponent()
         if edb_comp.IsNull():
