+++ src/pyedb/dotnet/edb_core/components.py
@@ -31,10 +31,10 @@ import warnings
 
 from pyedb.dotnet.clr_module import String
 from pyedb.dotnet.edb_core.definition.component_def import EDBComponentDef
-from pyedb.dotnet.edb_core.edb_data.components_data import EDBComponent
 from pyedb.dotnet.edb_core.edb_data.padstacks_data import EDBPadstackInstance
 from pyedb.dotnet.edb_core.edb_data.sources import Source, SourceType
 from pyedb.dotnet.edb_core.general import convert_py_list_to_net_list
+from pyedb.dotnet.edb_core.hierarchy.component import EDBComponent
 from pyedb.dotnet.edb_core.padstack import EdbPadstacks
 from pyedb.generic.general_methods import (
     _retry_ntimes,
