+++ src/pyedb/dotnet/edb_core/edb_data/ports.py
@@ -20,7 +20,7 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-from pyedb.dotnet.edb_core.edb_data.terminals import (
+from pyedb.dotnet.edb_core.terminal.terminal import (
     BundleTerminal,
     EdgeTerminal,
     PadstackInstanceTerminal,
