+++ src/pyedb/dotnet/edb_core/terminal/terminal.py
@@ -25,8 +25,14 @@ import re
 from pyedb.dotnet.edb_core.cell.layout_obj import Connectable
 from pyedb.dotnet.edb_core.edb_data.padstacks_data import EDBPadstackInstance
 from pyedb.dotnet.edb_core.edb_data.primitives_data import cast
-from pyedb.dotnet.edb_core.general import convert_py_list_to_net_list
-from pyedb.generic.general_methods import generate_unique_name, pyedb_function_handler
+from pyedb.dotnet.edb_core.terminal.bundle_terminal import BundleTerminal
+from pyedb.dotnet.edb_core.terminal.edge_terminal import EdgeTerminal
+from pyedb.dotnet.edb_core.terminal.padstack_instance_terminal import (
+    PadstackInstanceTerminal,
+)
+from pyedb.dotnet.edb_core.terminal.pingroup_terminal import PinGroupTerminal
+from pyedb.dotnet.edb_core.terminal.point_terminal import PointTerminal
+from pyedb.generic.general_methods import pyedb_function_handler
 
 
 class Terminal(Connectable):
@@ -432,224 +438,3 @@ class Terminal(Connectable):
                 pin_obj = pin
         if pin_obj:
             return EDBPadstackInstance(pin_obj, self._pedb)
-
-
-class EdgeTerminal(Terminal):
-    def __init__(self, pedb, edb_object):
-        super().__init__(pedb, edb_object)
-
-    @pyedb_function_handler()
-    def couple_ports(self, port):
-        """Create a bundle wave port.
-
-        Parameters
-        ----------
-        port : :class:`dotnet.edb_core.ports.WavePort`, :class:`dotnet.edb_core.ports.GapPort`, list, optional
-            Ports to be added.
-
-        Returns
-        -------
-        :class:`dotnet.edb_core.ports.BundleWavePort`
-
-        """
-        if not isinstance(port, (list, tuple)):
-            port = [port]
-        temp = [self._edb_object]
-        temp.extend([i._edb_object for i in port])
-        edb_list = convert_py_list_to_net_list(temp, self._edb.cell.terminal.Terminal)
-        _edb_bundle_terminal = self._edb.cell.terminal.BundleTerminal.Create(edb_list)
-        return self._pedb.ports[self.name]
-
-
-class BundleTerminal(Terminal):
-    """Manages bundle terminal properties.
-
-    Parameters
-    ----------
-    pedb : pyedb.edb.Edb
-        EDB object from the ``Edblib`` library.
-    edb_object : Ansys.Ansoft.Edb.Cell.Terminal.BundleTerminal
-        BundleTerminal instance from EDB.
-    """
-
-    def __init__(self, pedb, edb_object):
-        super().__init__(pedb, edb_object)
-
-    @property
-    def terminals(self):
-        """Get terminals belonging to this excitation."""
-        return [EdgeTerminal(self._pedb, i) for i in list(self._edb_object.GetTerminals())]
-
-    @property
-    def name(self):
-        return self.terminals[0].name
-
-    @pyedb_function_handler()
-    def decouple(self):
-        """Ungroup a bundle of terminals."""
-        return self._edb_object.Ungroup()
-
-
-class PadstackInstanceTerminal(Terminal):
-    """Manages bundle terminal properties."""
-
-    def __init__(self, pedb, edb_object):
-        super().__init__(pedb, edb_object)
-
-    @property
-    def position(self):
-        """Return terminal position.
-        Returns
-        -------
-        Position [x,y] : [float, float]
-        """
-        edb_padstack_instance = self._edb_object.GetParameters()
-        if edb_padstack_instance[0]:
-            return EDBPadstackInstance(edb_padstack_instance[1], self._pedb).position
-        return False
-
-    def create(self, padstack_instance, name=None, layer=None, is_ref=False):
-        """Create an edge terminal.
-
-        Parameters
-        ----------
-        prim_id : int
-            Primitive ID.
-        point_on_edge : list
-            Coordinate of the point to define the edge terminal.
-            The point must be on the target edge but not on the two
-            ends of the edge.
-        terminal_name : str, optional
-            Name of the terminal. The default is ``None``, in which case the
-            default name is assigned.
-        is_ref : bool, optional
-            Whether it is a reference terminal. The default is ``False``.
-
-        Returns
-        -------
-        Edb.Cell.Terminal.EdgeTerminal
-        """
-        if not name:
-            pin_name = padstack_instance._edb_object.GetName()
-            refdes = padstack_instance.component.refdes
-            name = "{}_{}".format(refdes, pin_name)
-            name = generate_unique_name(name)
-
-        if not layer:
-            layer = padstack_instance.start_layer
-
-        layer_obj = self._pedb.stackup.signal_layers[layer]
-
-        terminal = self._edb.cell.terminal.PadstackInstanceTerminal.Create(
-            self._pedb.active_layout,
-            padstack_instance.net.net_object,
-            name,
-            padstack_instance._edb_object,
-            layer_obj._edb_layer,
-            isRef=is_ref,
-        )
-        terminal = PadstackInstanceTerminal(self._pedb, terminal)
-
-        return terminal if not terminal.is_null else False
-
-
-class PointTerminal(Terminal):
-    """Manages point terminal properties."""
-
-    def __init__(self, pedb, edb_object=None):
-        super().__init__(pedb, edb_object)
-
-    @pyedb_function_handler
-    def create(self, name, net, location, layer, is_ref=False):
-        """Create a point terminal.
-
-        Parameters
-        ----------
-        name : str
-            Name of the terminal.
-        net : str
-            Name of the net.
-        location : list
-            Location of the terminal.
-        layer : str
-            Name of the layer.
-        is_ref : bool, optional
-            Whether it is a reference terminal.
-
-        Returns
-        -------
-        :class:`pyedb.dotnet.edb_core.edb_data.terminals.PointTerminal`
-        """
-        terminal = self._pedb.edb_api.cell.terminal.PointTerminal.Create(
-            self._pedb.active_layout,
-            self._pedb.nets[net].net_object,
-            name,
-            self._pedb.point_data(*location),
-            self._pedb.stackup[layer]._edb_layer,
-            is_ref,
-        )
-        terminal = PointTerminal(self._pedb, terminal)
-        return terminal if not terminal.is_null else False
-
-    @property
-    def location(self):
-        """Location of the terminal."""
-        layer = list(self._pedb.stackup.layers.values())[0]._edb_layer
-        _, point_data, _ = self._edb_object.GetParameters(None, layer)
-        return [point_data.X.ToDouble(), point_data.Y.ToDouble()]
-
-    @location.setter
-    def location(self, value):
-        layer = self.layer
-        self._edb_object.SetParameters(self._pedb.point_data(*value), layer)
-
-    @property
-    def layer(self):
-        """Get layer of the terminal."""
-        point_data = self._pedb.point_data(0, 0)
-        layer = list(self._pedb.stackup.layers.values())[0]._edb_layer
-        if self._edb_object.GetParameters(point_data, layer):
-            return layer
-
-    @layer.setter
-    def layer(self, value):
-        layer = self._pedb.stackup.layers[value]._edb_layer
-        point_data = self._pedb.point_data(*self.location)
-        self._edb_object.SetParameters(point_data, layer)
-
-
-class PinGroupTerminal(Terminal):
-    """Manages pin group terminal properties."""
-
-    def __init__(self, pedb, edb_object=None):
-        super().__init__(pedb, edb_object)
-
-    @pyedb_function_handler
-    def create(self, name, net_name, pin_group_name, is_ref=False):
-        """Create a pin group terminal.
-
-        Parameters
-        ----------
-        name : str
-            Name of the terminal.
-        net_name : str
-            Name of the net.
-        pin_group_name : str,
-            Name of the pin group.
-        is_ref : bool, optional
-            Whether it is a reference terminal. The default is ``False``.
-
-        Returns
-        -------
-        :class:`pyedb.dotnet.edb_core.edb_data.terminals.PinGroupTerminal`
-        """
-        net_obj = self._pedb.edb_api.cell.net.find_by_name(self._pedb.active_layout, net_name)
-        term = self._pedb.edb_api.cell.terminal.PinGroupTerminal.Create(
-            self._pedb.active_layout,
-            net_obj.api_object,
-            name,
-            self._pedb.siwave.pin_groups[pin_group_name]._edb_object,
-            is_ref,
-        )
-        term = PinGroupTerminal(self._pedb, term)
-        return term if not term.is_null else False
